+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Bugs Fixed
 
+- Fixed Resource Group option definition sharing static state between commands, which was causing test failures. Now each command can specify whether the resource group option is required without affecting other commands. [[#617](https://github.com/Azure/azure-mcp/issues/617)]
+
 ### Other Changes
 
 ## 0.4.0 (2025-07-15)